use WireModQ for mod(p) w/ k=1
Some comments.

Implementing mod(p^k) will require a label size of
k*Block, it is larger and needs to change
from_block and many other ops in HalfGate Mult.
So let us first implement mod(p). For large numbers,
it is quite enough with CRT. Not sure power(k) is
that efficient.